src/test/java/Kata009Test.java
import org.junit.Test;

import static org.junit.Assert.assertEquals;


public class Kata009Test {
    @Test
    public void testBerries() {
        Kata009 dictionary = new Kata009(new String[]{"cherry", "pineapple", "melon", "strawberry", "raspberry"});
        assertEquals("strawberry", dictionary.findMostSimilar("strawbery"));
        assertEquals("cherry", dictionary.findMostSimilar("berry"));
    }

    @Test
    public void testLanguages() {
        Kata009 dictionary = new Kata009(new String[]{"javascript", "java", "ruby", "php", "python", "coffeescript"});
        assertEquals("java", dictionary.findMostSimilar("heaven"));
        assertEquals("javascript", dictionary.findMostSimilar("javascript"));
    }

}